topics/pytorch/requirements.txt
absl-py @ file:///tmp/build/80754af9/absl-py_1607439979954/work
argon2-cffi==21.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
boto==2.49.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
cypari==2.4.1
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
docopt==0.6.2
entrypoints==0.3
et-xmlfile==1.0.1
findspark==1.4.2
fire==0.4.0
Flask==1.1.2
flatbuffers==1.12
future==0.18.2
FXrays==1.3.5
gast==0.3.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
gssapi==1.6.12
gunicorn==20.0.4
h5py==2.10.0
hdfs==2.6.0
idna==2.10
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
install==1.3.5
ipykernel==5.5.5
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
joblib==1.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.2
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
Keras==2.3.0
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keras2onnx==1.7.0
kiwisolver==1.3.1
krbcontext==0.10
lightgbm==3.2.1
Markdown @ file:///C:/ci/markdown_1605111171606/work
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
nbclient==0.5.4
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5
notebook==6.4.3
numpy==1.16.0
oauthlib==3.1.1
onnx==1.8.0
onnx-tf==1.6.0
onnx2keras==0.0.24
onnx2pytorch==0.4.1
onnxconverter-common==1.7.0
onnxoptimizer==0.2.6
onnxruntime==1.6.0
opencv-python==4.5.1.48
openpyxl==3.0.7
opt-einsum==3.3.0
packaging==21.0
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.0
plink==2.3.1
prometheus-client==0.11.0
prompt-toolkit==3.0.16
protobuf==3.13.0
py4j==0.10.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.8.0
pyparsing==2.4.7
pypmml==0.9.9
pypng==0.0.20
pyreadline==2.1
pyrsistent==0.18.0
pyspark==3.0.1
pysparkling==0.6.1
python-dateutil==2.8.1
pytz==2021.1
pywin32==301
pywinpty==1.1.4
PyYAML==5.4.1
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.11.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.8
scikit-learn==0.24.0
scipy==1.4.1
Send2Trash==1.8.0
six @ file:///C:/ci/six_1605187303045/work
sklearn==0.0
snappy==2.8
snappy-manifolds==1.1.2
snapy==1.0.2
spherogram==1.8.3
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.3.0
tensorflow-addons==0.13.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
tf2onnx==1.9.3
threadpoolctl==2.1.0
torch==1.10.1
torchvision==0.11.2
tornado==6.1
traitlets==4.3.3
typeguard==2.12.1
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.16.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
xgboost==1.4.2
xlrd==2.0.1
zipp @ file:///tmp/build/80754af9/zipp_1604001098328/work
